refactor(blorktools): html2canvas util
